New file Photo Structures was added. Slight implementations in ImagesListService .
